refactor(#1574): move C3 memo helpers to reader/bti_lookup_memo.rs (keep partition_lookup.rs under file-size threshold)
